@@ -17,6 +17,7 @@
 #include <climits>
 
 #include "exec/schema_scanner/schema_helper.h"
+#include "gutil/strings/substitute.h"
 #include "http/http_client.h"
 #include "runtime/runtime_state.h"
 #include "runtime/string_value.h"
@@ -70,7 +71,7 @@ Status SchemaLoadTrackingLogsScanner::fill_chunk(ChunkPtr* chunk) {
         auto& info = _result.loads[_cur_idx];
         for (const auto& [slot_id, index] : slot_id_to_index_map) {
             if (slot_id < 1 || slot_id > 4) {
-                return Status::InternalError(fmt::format("invalid slot id:$0", slot_id));
+                return Status::InternalError(strings::Substitute("invalid slot id: $0", slot_id));
             }
             ColumnPtr column = (*chunk)->get_column_by_slot_id(slot_id);
             switch (slot_id) {
