@@ -17,6 +17,7 @@
 #include <boost/algorithm/string/predicate.hpp>
 
 #include "fmt/format.h"
+#include "gutil/strings/substitute.h"
 #include "http/http_client.h"
 #include "util/defer_op.h"
 #include "util/md5.h"
@@ -47,7 +48,8 @@ Status DownloadUtil::download(const std::string& url, const std::string& tmp_fil
         auto res = fwrite(data, length, 1, fp);
         if (res != 1) {
             LOG(ERROR) << fmt::format("fail to write data to file {}, error={}", tmp_file, ferror(fp));
-            status = Status::InternalError(fmt::format("file to write data when downloading file from {}" + url));
+            status =
+                    Status::InternalError(strings::Substitute("file to write data when downloading file from $0", url));
             return false;
         }
         return true;
